+++ sormas-ui/src/main/java/de/symeda/sormas/ui/utils/components/sidecomponent/event/sidecomponentfield/SideComponentFieldEditEventListener.java
@@ -1,4 +1,4 @@
-package de.symeda.sormas.ui.utils.components.sidecomponent.event;
+package de.symeda.sormas.ui.utils.components.sidecomponent.event.sidecomponentfield;
 
 import java.lang.reflect.Method;
